@@ -14,7 +14,7 @@ dependencies:
   sqflite: ^2.0.2+1
   path: ^1.8.0
 
-  objectbox: ^1.5.0
+  objectbox: ^2.0.0
   objectbox_flutter_libs: any
 
   shared_preferences: ^2.0.6
@@ -25,8 +25,6 @@ dependencies:
 
   provider: ^6.0.0
 
-  modal_bottom_sheet: ^2.0.0
-
   url_launcher: ^6.0.6
   share_plus: ^6.3.0
 
@@ -35,7 +33,7 @@ dependencies:
 
   flutter_cached_pdfview: ^0.4.1
 
-  flutter_svg: ^1.0.0
+  flutter_svg: ^2.0.4
   flutter_staggered_grid_view: ^0.6.1
 
   adaptive_dialog: ^1.6.4
@@ -56,13 +54,13 @@ dependencies:
 
   font_awesome_flutter: ^10.1.0
 
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.3
 
   cupertino_icons: ^1.0.3
 
-  sentry_flutter: ^6.7.0
+  sentry_flutter: ^7.4.0
 
-  package_info_plus: ^1.3.0
+  package_info_plus: ^3.0.3
   web_socket_channel: ^2.2.0
   flutter_slidable: ^2.0.0
   firebase_core: ^2.4.1
@@ -73,6 +71,9 @@ dependencies:
   flat_buffers: ^2.0.5
   collection: ^1.16.0
   http: ^0.13.5
+
+  presentation:
+    path: ../presentation
 
 dev_dependencies:
   build_runner: ^2.0.4
